pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docrestore"
version = "0.1.0"
description = "处理文档屏幕拍摄照片，还原为原文格式的 markdown 文档"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1",
    "aiosqlite>=0.20",
    "fastapi>=0.115",
    "litellm>=1.50",
    "Pillow>=10.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "mypy>=1.13",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "types-aiofiles>=24.1",
]
ocr = [
    # torch/vllm/transformers/tokenizers 由 scripts/setup_deepseek_ocr.sh 按特定顺序安装
    # （vllm 和 transformers 有版本冲突，需强制降级）
    "einops",
    "easydict",
    "addict",
    "numpy",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/docrestore"]

[tool.ruff]
src = ["backend"]
target-version = "py311"
exclude = ["third_party"]

[tool.ruff.lint]
select = [
    "E", "F", "W",
    "B",       # bugbear
    "C4",      # comprehensions
    "C90",     # mccabe
    "UP",      # pyupgrade
    "T20",     # print
    "ARG",     # unused arguments
    "RET",     # return
    "S",       # bandit
    "ASYNC",   # async
    "SIM105",  # contextlib.suppress
    "PT",      # pytest
]
ignore = [
    "RUF001", "RUF002", "RUF003",  # unicode 字符
    "S101",   # assert 在测试中使用
    "ASYNC240",  # pathlib 在 async 中使用（快速本地 IO 无需异步化）
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG001", "ARG002", "T201"]
"scripts/**/*.py" = ["T201"]

[tool.mypy]
strict = true
python_version = "3.11"
mypy_path = "backend"
packages = ["docrestore"]
explicit_package_bases = true
plugins = []
exclude = ["third_party"]

[[tool.mypy.overrides]]
module = [
    "vllm", "vllm.*",
    "litellm.*", "deepseek_ocr2.*",
    "torch.*", "torchvision.*", "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    # 第三方 C 扩展（protobuf/grpc）的 DeprecationWarning，无法在本项目修复
    "ignore::DeprecationWarning:importlib._bootstrap",
    # litellm 内部未 await 的协程
    "ignore::RuntimeWarning:litellm",
    # pytest stash 清理时 AsyncMock 协程未 await（pytest 内部机制）
    "ignore::RuntimeWarning:_pytest.stash",
]
